app/api/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
psycopg2-binary==2.9.6
